+++ R/get_predicted.R
@@ -1,0 +1,19 @@
+#' Predicted values
+#'
+#' Returns values predicted by a model (i.e., fitted values).
+#'
+#' @inheritParams get_residuals
+#'
+#' @examples
+#' x <- lm(Sepal.Length ~ Petal.Width * Species, data=iris)
+#'
+#' get_predicted(x)
+#' @export
+get_predicted <- function(x, ...){
+  UseMethod("get_predicted")
+}
+
+#' @export
+get_predicted.default <- function(x, ...){
+  stats::fitted(x)
+}